% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_process.R
\name{RunGame}
\alias{RunGame}
\title{Title}
\usage{
RunGame(starting_deck, decisions1, decisions2, starting_player)
}
\arguments{
\item{starting_player}{The game state is:
deck (card to draw)
hand1
hand2
stack1
stack2
board}
}
\value{

}
\description{
Title
}
